refactor: extract shared monitor loop and report into monitor_utils
Move duplicated build_report() and fetch-filter-report loop from
community-monitor.py and native-sources-monitor.py into shared
run_monitor() and build_report() in lib/monitor_utils.py.

Addresses CodeFactor duplication finding in lines 405-452/406-453.
