+++ .github/scripts/community-monitor.py
@@ -21,19 +21,14 @@ import json
 import os
 import re
 import sys
-import urllib.error
 import urllib.parse
 import urllib.request
 from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).parent))
 from lib.monitor_utils import (
-    build_doc_keywords,
-    entry_fingerprint,
     fetch_text,
-    is_covered,
-    load_state,
-    save_state,
+    run_monitor,
 )
 
 # ---------------------------------------------------------------------------
@@ -298,71 +293,6 @@ def fetch_and_extract_source(source: dict[str, str]) -> list[dict[str, str]]:
 
 
 # ---------------------------------------------------------------------------
-# Report generation
-# ---------------------------------------------------------------------------
-
-def build_report(
-    source_results: list[dict],
-) -> tuple[str, bool]:
-    """Build a markdown report of uncovered community content.
-
-    Returns (report_text, has_new_content).
-    """
-    lines: list[str] = [
-        "## Community Monitor Report",
-        "",
-        f"Sources checked: **{len(source_results)}**",
-        "",
-    ]
-
-    total_new = 0
-
-    for result in source_results:
-        name = result["name"]
-        new_entries = result["new_entries"]
-        total_new += len(new_entries)
-
-        lines += [
-            f"### {name}",
-            "",
-            f"- Source: {result['description']}",
-            f"- Entries fetched: {result['total']}",
-            f"- New uncovered: {len(new_entries)}",
-            "",
-        ]
-
-        if new_entries:
-            lines.append("| Entry | Section | Description |")
-            lines.append("|-------|---------|-------------|")
-            for entry in new_entries[:30]:  # Cap at 30 per source
-                name_col = entry.get("name", "")[:60]
-                heading = entry.get("heading", "")[:40]
-                desc = entry.get("description", "")[:80]
-                # Escape pipe chars in table cells
-                name_col = name_col.replace("|", "\\|")
-                heading = heading.replace("|", "\\|")
-                desc = desc.replace("|", "\\|")
-                lines.append(f"| {name_col} | {heading} | {desc} |")
-            lines.append("")
-
-            if len(new_entries) > 30:
-                lines.append(
-                    f"_... and {len(new_entries) - 30} more entries._"
-                )
-                lines.append("")
-
-    lines += [
-        "---",
-        f"Total new uncovered entries: **{total_new}**",
-        "",
-        "_Generated by `.github/scripts/community-monitor.py`_",
-        "",
-    ]
-
-    return "\n".join(lines), total_new > 0
-
-
-# ---------------------------------------------------------------------------
 # Main
 # ---------------------------------------------------------------------------
 
@@ -387,65 +317,14 @@ def main() -> None:
             f"ERROR: Community docs dir does not exist: {args.community_docs_dir}"
         )
 
-    # Build keyword index from existing community docs
-    doc_keywords = build_doc_keywords(args.community_docs_dir)
-    print(f"Doc keywords indexed: {len(doc_keywords)}", file=sys.stderr)
-
-    # Load previous state
-    state = load_state(args.state_file)
-
-    source_results: list[dict] = []
-
-    for source in SOURCES:
-        name = source["name"]
-        print(f"Fetching {name}...", file=sys.stderr)
-
-        try:
-            entries = fetch_and_extract_source(source)
-        except Exception as e:
-            print(f"WARNING: Failed to fetch {name}: {e}", file=sys.stderr)
-            source_results.append({
-                "name": name,
-                "description": source["description"],
-                "total": 0,
-                "new_entries": [],
-            })
-            continue
-
-        print(f"  Extracted {len(entries)} entries", file=sys.stderr)
-
-        # Filter to entries not previously seen and not covered by docs
-        seen = set(state.get(name, []))
-        new_entries: list[dict[str, str]] = []
-
-        for entry in entries:
-            fp = entry_fingerprint(entry)
-            if fp in seen:
-                continue
-            if not is_covered(entry, doc_keywords):
-                new_entries.append(entry)
-
-        print(f"  New uncovered: {len(new_entries)}", file=sys.stderr)
-
-        # Update state with all current fingerprints
-        state[name] = [entry_fingerprint(e) for e in entries]
-
-        source_results.append({
-            "name": name,
-            "description": source["description"],
-            "total": len(entries),
-            "new_entries": new_entries,
-        })
-
-    # Save updated state
-    save_state(args.state_file, state)
-    print(f"State saved to {args.state_file}", file=sys.stderr)
-
-    # Build and output report
-    report, has_new = build_report(source_results)
-    print(report)
-
-    sys.exit(1 if has_new else 0)
+    run_monitor(
+        sources=SOURCES,
+        fetch_fn=fetch_and_extract_source,
+        docs_dir=args.community_docs_dir,
+        state_file=args.state_file,
+        report_title="Community Monitor Report",
+        script_name="community-monitor.py",
+    )
 
 
 if __name__ == "__main__":
