+++ .github/scripts/lib/monitor_utils.py
@@ -8,7 +8,9 @@ community-monitor, and native-sources-monitor.
 import hashlib
 import json
 import re
+import sys
 import urllib.request
+from collections.abc import Callable
 from pathlib import Path
 
 # Noise words that match too broadly across docs
@@ -76,6 +78,141 @@ def save_state(state_file: Path, state: dict[str, list[str]]) -> None:
     """Save seen entry fingerprints."""
     state_file.parent.mkdir(parents=True, exist_ok=True)
     state_file.write_text(json.dumps(state, indent=2) + "\n", encoding="utf-8")
+
+
+def build_report(
+    source_results: list[dict],
+    report_title: str,
+    script_name: str,
+) -> tuple[str, bool]:
+    """Build a markdown report of uncovered content.
+
+    Args:
+        source_results: List of per-source result dicts.
+        report_title: H2 heading for the report.
+        script_name: Script filename shown in the footer.
+
+    Returns:
+        (report_text, has_new_content).
+    """
+    lines: list[str] = [
+        f"## {report_title}",
+        "",
+        f"Sources checked: **{len(source_results)}**",
+        "",
+    ]
+
+    total_new = 0
+
+    for result in source_results:
+        name = result["name"]
+        new_entries = result["new_entries"]
+        total_new += len(new_entries)
+
+        lines += [
+            f"### {name}",
+            "",
+            f"- Source: {result['description']}",
+            f"- Entries fetched: {result['total']}",
+            f"- New uncovered: {len(new_entries)}",
+            "",
+        ]
+
+        if new_entries:
+            lines.append("| Entry | Section | Description |")
+            lines.append("|-------|---------|-------------|")
+            for entry in new_entries[:30]:
+                name_col = entry.get("name", "")[:60].replace("|", "\\|")
+                heading = entry.get("heading", "")[:40].replace("|", "\\|")
+                desc = entry.get("description", "")[:80].replace("|", "\\|")
+                lines.append(f"| {name_col} | {heading} | {desc} |")
+            lines.append("")
+
+            if len(new_entries) > 30:
+                lines.append(f"_... and {len(new_entries) - 30} more entries._")
+                lines.append("")
+
+    lines += [
+        "---",
+        f"Total new uncovered entries: **{total_new}**",
+        "",
+        f"_Generated by `.github/scripts/{script_name}`_",
+        "",
+    ]
+
+    return "\n".join(lines), total_new > 0
+
+
+def run_monitor(
+    sources: list[dict[str, str]],
+    fetch_fn: Callable[[dict[str, str]], list[dict[str, str]]],
+    docs_dir: Path,
+    state_file: Path,
+    report_title: str,
+    script_name: str,
+) -> None:
+    """Shared monitor loop: fetch sources, filter, report, exit.
+
+    Args:
+        sources: Source definitions to poll.
+        fetch_fn: Callable that fetches and extracts entries for a source.
+        docs_dir: Directory of existing docs for coverage checking.
+        state_file: Path to JSON state file.
+        report_title: H2 heading for the report.
+        script_name: Script filename for the report footer.
+    """
+    doc_keywords = build_doc_keywords(docs_dir)
+    print(f"Doc keywords indexed: {len(doc_keywords)}", file=sys.stderr)
+
+    state = load_state(state_file)
+    source_results: list[dict] = []
+
+    for source in sources:
+        name = source["name"]
+        print(f"Fetching {name}...", file=sys.stderr)
+
+        try:
+            entries = fetch_fn(source)
+        except Exception as e:
+            print(f"WARNING: Failed to fetch {name}: {e}", file=sys.stderr)
+            source_results.append({
+                "name": name,
+                "description": source["description"],
+                "total": 0,
+                "new_entries": [],
+            })
+            continue
+
+        print(f"  Extracted {len(entries)} entries", file=sys.stderr)
+
+        seen = set(state.get(name, []))
+        new_entries: list[dict[str, str]] = []
+
+        for entry in entries:
+            fp = entry_fingerprint(entry)
+            if fp in seen:
+                continue
+            if not is_covered(entry, doc_keywords):
+                new_entries.append(entry)
+
+        print(f"  New uncovered: {len(new_entries)}", file=sys.stderr)
+
+        state[name] = [entry_fingerprint(e) for e in entries]
+
+        source_results.append({
+            "name": name,
+            "description": source["description"],
+            "total": len(entries),
+            "new_entries": new_entries,
+        })
+
+    save_state(state_file, state)
+    print(f"State saved to {state_file}", file=sys.stderr)
+
+    report, has_new = build_report(source_results, report_title, script_name)
+    print(report)
+
+    sys.exit(1 if has_new else 0)
 
 
 def fetch_text(
